Move tabular chunked view component to datasets, its not a visualization
